requirements.txt
django
pillow
django-ckeditor 